+++ ft_strstr.c
@@ -6,7 +6,7 @@
 /*   By: labdello <labdello@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/06 16:23:39 by labdello          #+#    #+#             */
-/*   Updated: 2024/04/06 22:40:05 by labdello         ###   ########.fr       */
+/*   Updated: 2024/04/07 13:01:32 by labdello         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strstr(char *str, char *needle)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (ft_strlen(needle) < 1)
